requirements.txt
praw
python-dotenv
google-generativeai>=0.5.0
